+++ src/Modules/OrchardCore.Commerce/Services/CardPaymentService.cs
@@ -1,6 +1,8 @@
 using Microsoft.AspNetCore.DataProtection;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Localization;
 using Microsoft.Extensions.Logging;
+using Newtonsoft.Json.Linq;
 using OrchardCore.Commerce.Abstractions;
 using OrchardCore.Commerce.Constants;
 using OrchardCore.Commerce.Extensions;
@@ -14,6 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using static OrchardCore.Commerce.Constants.ContentTypes;
 
 namespace OrchardCore.Commerce.Services;
 
@@ -29,6 +32,8 @@ public class CardPaymentService : ICardPaymentService
     private readonly IDataProtectionProvider _dataProtectionProvider;
     private readonly ILogger<CardPaymentService> _logger;
     private readonly IStringLocalizer T;
+    private readonly Lazy<IUserService> _userServiceLazy;
+    private readonly IHttpContextAccessor _hca;
 
     // We need to use that many this cannot be avoided.
 #pragma warning disable S107 // Methods should not have too many parameters
@@ -41,7 +46,9 @@ public class CardPaymentService : ICardPaymentService
         ISiteService siteService,
         IDataProtectionProvider dataProtectionProvider,
         ILogger<CardPaymentService> logger,
-        IStringLocalizer<CardPaymentService> stringLocalizer)
+        IStringLocalizer<CardPaymentService> stringLocalizer,
+        Lazy<IUserService> userServiceLazy,
+        IHttpContextAccessor hca)
 #pragma warning restore S107 // Methods should not have too many parameters
     {
         _paymentIntentService = new PaymentIntentService();
@@ -53,6 +60,8 @@ public class CardPaymentService : ICardPaymentService
         _siteService = siteService;
         _dataProtectionProvider = dataProtectionProvider;
         _logger = logger;
+        _userServiceLazy = userServiceLazy;
+        _hca = hca;
         T = stringLocalizer;
     }
 
@@ -109,11 +118,6 @@ public class CardPaymentService : ICardPaymentService
                 ConfirmationMethod = "manual",
                 Confirm = true,
                 PaymentMethod = paymentMethodId,
-
-                // If shipping is implemented, it needs to be added here too.
-                // https://github.com/OrchardCMS/OrchardCore.Commerce/issues/4
-                // Shipping =
-                // ReceiptEmail = viewModel.Email,
             };
 
             paymentIntent = await _paymentIntentService.CreateAsync(paymentIntentOptions, requestOptions);
@@ -143,10 +147,6 @@ public class CardPaymentService : ICardPaymentService
 
         order.DisplayText = T["Order {0}", orderId];
 
-        // To-do when other parts of the checkout is implemented (notes).
-        // https://github.com/OrchardCMS/OrchardCore.Commerce/issues/4
-        // order.Alter<HtmlBodyPart>(htmlBodyPart => htmlBodyPart.
-
         IList<OrderLineItem> lineItems = new List<OrderLineItem>();
 
         // This needs to be done separately because it's async: "Avoid using async lambda when delegate type returns
@@ -172,13 +172,29 @@ public class CardPaymentService : ICardPaymentService
 
             orderPart.OrderId.Text = orderId.ToString();
             orderPart.Status.Text = OrderStatuses.Ordered;
-
-            // To-do when shipping is implemented. https://github.com/OrchardCMS/OrchardCore.Commerce/issues/4
-            // oderPartContent.BillingAddress
-            // oderPartContent.ShippingAddress
         });
 
         await _contentManager.CreateAsync(order);
+
+        // Saving addresses.
+        var userService = _userServiceLazy.Value;
+        var orderPart = order.As<OrderPart>();
+
+        if (await userService.GetCurrentFullUserAsync(_hca) is { } user)
+        {
+            var isSame = orderPart.BillingAndShippingAddressesMatch.Value;
+
+            await userService.AlterUserSettingAsync(user, UserAddresses, contentItem =>
+            {
+                var part = contentItem.ContainsKey(nameof(UserAddressesPart))
+                    ? contentItem[nameof(UserAddressesPart)].ToObject<UserAddressesPart>()!
+                    : new UserAddressesPart();
+
+                part.BillingAndShippingAddressesMatch.Value = isSame;
+                contentItem[nameof(UserAddressesPart)] = JToken.FromObject(part);
+                return contentItem;
+            });
+        }
 
         currentShoppingCart.Items.Clear();
